Struct reaper_medium::OwnedPcmSource

#[repr(transparent)]
pub struct OwnedPcmSource(_);

Owned PCM source.

This PCM source automatically destroys the associated C++ PCM_source when dropped.

Implementations

impl OwnedPcmSource

pub unsafe fn from_raw(raw: PcmSource) -> Self

Takes ownership of the given source.

Safety

You must guarantee that the given source is currently owner-less, otherwise double-free or use-after-free can occur.

pub unsafe fn leak(self) -> PcmSource

Returns the inner pointer without destroying the source.

Safety

You can run into a memory leak or crash if you don’t manage the lifetime of the returned source correctly.