Struct reaper_medium::PcmSourceTransfer

#[repr(transparent)]
pub struct PcmSourceTransfer(_);

PCM source transfer.

Implementations

impl PcmSourceTransfer

pub fn as_ptr(&self) -> NonNull<PCM_source_transfer_t>

Returns the pointer to this source transfer.

pub fn time_s(&self) -> PositionInSeconds

Returns the start time of the block.

pub fn set_time_s(&mut self, time: PositionInSeconds)

Sets the start time of the block.

pub fn absolute_time_s(&self) -> PositionInSeconds

Returns the absolute start time of the block.

pub fn set_absolute_time_s(&mut self, time: PositionInSeconds)

Sets the absolute start time of the block.

pub fn samples_out(&self) -> i32

Returns the number of sample(pair)s actually rendered.

pub unsafe fn set_samples_out(&mut self, number: i32)

Sets the number of sample(pair)s actually rendered.

Safety

TODO-high-unstable

pub fn length(&self) -> i32

Returns the desired number of sample(pair)s to be filled.

Safety

TODO-high-unstable

pub unsafe fn set_length(&mut self, length: i32)

Sets the desired number of sample(pair)s to be filled.

Safety

TODO-high-unstable

pub fn nch(&self) -> i32

Returns the desired number of channels.

pub unsafe fn set_nch(&mut self, nch: i32)

Sets the desired number of channels.

Safety

TODO-high-unstable

pub fn samples(&self) -> *mut f64

Returns the sample(pair)s to be rendered.

pub unsafe fn samples_as_slice(&self) -> &[f64]

Returns the samples as read-only slice.

Safety

If the length or the samples are set incorrectly, this results in undefined behavior.

TODO-high-unstable

pub unsafe fn samples_as_mut_slice(&mut self) -> &mut [f64]

Returns the samples as mutable slice.

Safety

If the length or the samples are set incorrectly, this results in undefined behavior.

TODO-high-unstable

pub unsafe fn set_samples(&mut self, samples: *mut f64)

Sets the sample(pair)s to be rendered.

Safety

TODO-high-unstable

pub fn sample_rate(&self) -> Hz

Returns the desired output sample rate.

pub fn set_sample_rate(&mut self, rate: Hz)

Sets the desired output sample rate.

pub fn midi_event_list_mut(&mut self) -> Option<&mut BorrowedMidiEventList>

Returns the list of MIDI events to be filled.

pub fn set_midi_event_list(&mut self, list: &BorrowedMidiEventList)

Sets the list of MIDI events to be filled. TODO-high This is bad modeling. Passing a reference and saving it.

pub fn force_bpm(&self) -> Bpm

pub fn set_force_bpm(&mut self, force_bpm: Bpm)